feat: update integer type mappings across multiple languages to use 'int' and 'int32'

+++ internal/emitter/lang/csharp.go
@@ -31,7 +31,7 @@ func (a *CSharpAdapter) MapType(veldType string) (string, []string, error) {
 
 	builtins := map[string]string{
 		"string":   "string",
-		"int":      "long",
+		"int":      "int",
 		"float":    "double",
 		"decimal":  "decimal",
 		"bool":     "bool",
